Simplify functests job
At one point we had an array of functest jobs; we were testing
building on trusty, bionic, maybe centos, with python2 and python3.
The only thing left of all these combos these days is "bionic-python3"
(that's where this naming comes from).

We can remove the levels of abstraction now and just have the one job
to avoid confusion.

Change-Id: Id37d62a17f9b7f6dc6dc35585c29eddd435ce913