Switch to release-notes-jobs-python3
The release-notes-jobs project-template ceased publishing release
notes in the tag pipeline in 2018 when
https://review.opendev.org/622430 merged. Projects were expected to
switch their master branches to release-notes-jobs-python3 instead
around that time, but DIB seems to have missed the boat. Update to
the modern one so that we'll go back to updating our published
release notes every time a new release is tagged.

Change-Id: I9268811438d690d3f945b5d651b8b2ff6220bb96