Switch from cp to mv when baking data into the raw disk image
When building diskimages and baking in large amounts of data, the data can exist
on the diskimage building machine essentially twice. When there is limited disk
space available on the machine, this becomes a problem.

The proposed change removes the cp and rm approach in favour of a mv
approach. This should remove the need for the data to exist twice on the
machine and reduce the chances for the machine to run out of disk space.