Replace OpenSUSE Leap 15.3 to OpenSUSE Leap 15.5
OpenSUSE Leap 15.3 is out of support since 2022. This patch remove the
support of OpenSUSE Leap 15.3 and replace it with the OpenSUSE Leap
15.5.
Add the grub2-x86_64-efi package in the pkg-map of the bootloader
element to allow EFI support.
Replace version 15.3 to 15.5 in every required location.
Add dracut-mkinitrd-deprecated package in package-installs.yaml of
the opensuse element to support the usage of the deprecated mkinitrd
command from the zypper element.

Change-Id: Ie903ffcf229d8e32374c4a1346b5295990c83ff4