Write fedora download redirect info to stderr
This will allow us to debug potentially bad fedora download backend
locations that don't provide valid subreleases in file paths. To do this
we ask curl to write headers (-i) and then dump the headers (-D) to
stderr. This should keep the debug info out of the parsing path for
subrelease info.

Change-Id: I07dd6ec4087191e1c688968be2fe1a1423a26b9e