Fetch compatibile dnf download command in functest setup
The dnf-plugins-core repo updates its download command to use a
dnf.utils method that is not present in the dnf version installed by
Debian packages. Update the fetch of dnf-plugins-core to use the last
version of the download plugin that is compatible with the dnf package
in Debian.

Note that we don't use the bookworm dnf-plugins-core package to address
this because dnf-plugins-core specifies that it breaks and replaces
zypper. This occurs due to the collision of a single manpage filename.
OpenSUSE addresses this by moving one of the manpages to a different
location allowing zypper and dnf to be installed together on SUSE
distros. Hopefully Debian will manage to do similar eventually.

Change-Id: I3a029340f9635f54570e48a68a9e2c1b1e30a76f