Remove vm element from container based image
The vm element sets up disk partitioning and bootloader installation in
dib built images. The container based elements are unique among distro
elements in that they have a hard dependency on the vm element.  Some
dib users may not need a complete disk image with partition tables and
bootloaders; we remove the vm element to give them that flexibility.

Change-Id: I43bb19415eb0b9c62e44c131497c28ae260369e7