Disable Stream-8, non-vote FC 37, and assume yes on debian functest
CentOS Stream 8 and Fedora 37 are EOL and their jobs are just
failing.

Disabled the Centos Stream 8 job, and marked the Fedora Core 37 job
as non-voting fo the time being so we are encouraged to fix by adding
a FC 38 or FC 39 job.

Also removing CS8 and Fedora jobs from dib-functests, and
setting the debian functional build to assume yes to try and
minimize prompting on package upgrades.

Change-Id: Iba7a52a7882de4293f87b98a61bda04faf07cbe8