Fix manifest element with non-root user
When building an image with the manifests element as a non-root user,
the image can fail to build. This affects the Ironic Python Agent (IPA)
image.

The issue was permission denied when copying the manifests from the
image build directory due to them being owned by the root user.

This change fixes the issue by copying the files using sudo, and
changing their ownership later.

Change-Id: I5fcdd9d47f97f32a5b4b8246e8b57ead41b0bdd9
Closes-Bug: #2069956